requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
odfpy==1.4.1
python-dotenv==1.0.0
playwright==1.40.0
requests==2.31.0
streamlit==1.31.1
plotly==5.18.0
matplotlib==3.8.2
Pillow==10.2.0
